+++ src/auth/auth.module.ts
@@ -3,7 +3,7 @@ import { AuthService } from './auth.service';
 import { PrismaModule } from '../prisma/prisma.module';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { JwtStrategy } from '../common/strategy/jwt.strategy';
+import { JwtStrategy } from './jwt/jwt.strategy';
 import { CryptoModule } from '../crypto/crypto.module';
 
 @Module({
